refactor: move componentes from stacks folder to components, rename files, move small components to the same file
