--- packages/app/app/stacks/components/stack-modal/components/StackDetail.tsx
@@ -1,17 +1,0 @@
-import { BodyText } from "@/ui";
-import { ReactNode } from "react";
-
-export const StackDetail = ({
-  title,
-  children,
-}: {
-  title: string;
-  children: ReactNode;
-}) => (
-  <div className="space-y-1">
-    <BodyText size={1} className="text-em-low">
-      {title}
-    </BodyText>
-    <BodyText size={1}>{children}</BodyText>
-  </div>
-);
