+++ packages/app/app/stacks/page.tsx
@@ -3,7 +3,7 @@
 import { Tab } from "@headlessui/react";
 import { EmptyState } from "@/app/stacks/empty-state";
 import { ButtonLink, HeadingText } from "@/ui";
-import { StacksTable } from "@/app/stacks/components/StacksTable";
+import { StacksTable } from "@/components/StacksTable";
 import { currentTimestampInSeconds } from "@/utils/time";
 import { Order } from "@/app/models/order";
